# docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shopping Mall API",
    "description": "Spring Boot 网上购物商城接口文档（可直接导入 Apifox）。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "tags": [
    { "name": "Auth", "description": "注册与登录" },
    { "name": "Product", "description": "商品" },
    { "name": "Cart", "description": "购物车" },
    { "name": "Order", "description": "订单" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "用户注册",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRequest" },
              "example": { "username": "test", "password": "123456" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthSuccessResponse" }
              }
            }
          },
          "400": {
            "description": "参数或业务校验失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "用户登录",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRequest" },
              "example": { "username": "test", "password": "123456" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功（会写入 Session Cookie）",
            "headers": {
              "Set-Cookie": {
                "description": "JSESSIONID",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthSuccessResponse" }
              }
            }
          },
          "400": {
            "description": "参数或业务校验失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Product"],
        "summary": "获取商品列表",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" }
                }
              }
            }
          }
        }
      }
    },
    "/api/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "查看购物车",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartResponse" }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/cart/add": {
      "post": {
        "tags": ["Cart"],
        "summary": "加入购物车",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddCartRequest" },
              "example": { "productId": 1, "quantity": 2 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "加入成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartAddResponse" }
              }
            }
          },
          "400": {
            "description": "参数或业务失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "tags": ["Order"],
        "summary": "创建订单（进入待付款并扣库存）",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderCreateResponse" }
              }
            }
          },
          "400": {
            "description": "业务失败（如库存不足、购物车为空）",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Order"],
        "summary": "查询当前用户订单",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}/pay": {
      "post": {
        "tags": ["Order"],
        "summary": "支付订单",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "支付成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderActionResponse" }
              }
            }
          },
          "400": {
            "description": "业务失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}/refund": {
      "post": {
        "tags": ["Order"],
        "summary": "退款订单（回退库存）",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "退款成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderActionResponse" }
              }
            }
          },
          "400": {
            "description": "业务失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    },
    "schemas": {
      "AuthRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "AuthSuccessResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "登录成功" },
          "userId": { "type": "integer", "format": "int64", "example": 1 },
          "username": { "type": "string", "example": "test" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "请先登录" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number", "format": "double" },
          "stock": { "type": "integer", "format": "int32" }
        }
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "product": { "$ref": "#/components/schemas/Product" },
          "quantity": { "type": "integer", "format": "int32" }
        }
      },
      "CartResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CartItem" }
          },
          "totalPrice": { "type": "number", "format": "double" }
        }
      },
      "CartAddResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/CartResponse" },
          {
            "type": "object",
            "properties": {
              "message": { "type": "string", "example": "加入购物车成功" }
            }
          }
        ]
      },
      "AddCartRequest": {
        "type": "object",
        "required": ["productId", "quantity"],
        "properties": {
          "productId": { "type": "integer", "format": "int64", "example": 1 },
          "quantity": { "type": "integer", "format": "int32", "minimum": 1, "example": 2 }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userId": { "type": "integer", "format": "int64" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CartItem" }
          },
          "totalPrice": { "type": "number", "format": "double" },
          "createdAt": { "type": "string", "format": "date-time" },
          "status": {
            "type": "string",
            "enum": ["PENDING_PAYMENT", "PAID", "REFUNDED"]
          }
        }
      },
      "OrderCreateResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "下单成功，订单进入待付款状态" },
          "order": { "$ref": "#/components/schemas/Order" }
        }
      },
      "OrderActionResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "支付成功" },
          "order": { "$ref": "#/components/schemas/Order" }
        }
      }
    }
  }
}